refactor(assistant): 拆分终态与后台任务生命周期支持
