+++ pinvou3-app/src-tauri/src/features/assistant/mod.rs
@@ -1,6 +1,7 @@
 pub(crate) mod audit;
 pub mod engine;
 pub(crate) mod engine_pool;
+mod engine_support;
 pub(crate) mod harness;
 pub mod platform;
 pub(crate) mod runtime_model;
