+++ pinvou3-app/src/index.html
@@ -158,6 +158,8 @@
   <script src="%BASE_URL%platform/web/bootstrap.js"></script>
   <script src="%BASE_URL%platform/web/host-file-picker.js"></script>
   <script src="%BASE_URL%features/attachments/attachment-drop-controller.js"></script>
+  <script src="%BASE_URL%shared/bridge-messages.js"></script>
+  <script src="%BASE_URL%platform/web/bridge/turn-terminal.js"></script>
   <script src="%BASE_URL%platform/web/bridge.js"></script>
   <script src="%BASE_URL%platform/web/bridge/domain-adapter.js"></script>
   <script src="%BASE_URL%platform/tauri/bridge/artifact-tracker.js"></script>
