+++ pinvou3-app/src/platform/tauri/bridge.js
@@ -392,7 +392,6 @@
       shellTaskFinished: code => `[Task finished, exit code: ${code}]`,
       skillContentHidden: "(Skill loaded, content hidden)",
       desktopDoneSyncPending: "⚠️ The conversation finished on desktop, but the authoritative record has not synced yet; you can retry once reconnected.",
-      shellCleanupFailed: "⚠️ Some background tasks could not be stopped. You can stop them individually from the background task list.",
       sessionSyncingTurn: "This chat is still syncing a turn completed elsewhere. Please try again shortly.",
       targetSessionMissing: "Target chat does not exist",
       replyContentEmpty: "Reply content is empty",
@@ -479,7 +478,6 @@
       shellTaskFinished: code => `[タスク終了、終了コード: ${code}]`,
       skillContentHidden: "（スキルを読み込みました。内容は非表示です）",
       desktopDoneSyncPending: "⚠️ 会話はデスクトップ側で完了しましたが、権威レコードはまだ同期されていません。接続回復後に再試行できます。",
-      shellCleanupFailed: "⚠️ 一部のバックグラウンドタスクを停止できませんでした。バックグラウンドタスク一覧から個別に停止できます。",
       sessionSyncingTurn: "このチャットは別端末で完了したターンを同期中です。しばらくしてから再試行してください",
       targetSessionMissing: "対象のチャットが存在しません",
       replyContentEmpty: "返信内容が空です",
@@ -566,7 +564,6 @@
       shellTaskFinished: code => `[任务已结束，退出码: ${code}]`,
       skillContentHidden: "（技能已加载，内容不展示）",
       desktopDoneSyncPending: "⚠️ 对话已在桌面端完成，但权威记录暂未同步；恢复连接后可重试。",
-      shellCleanupFailed: "⚠️ 部分后台任务未能停止，可在后台任务列表中逐个停止。",
       sessionSyncingTurn: "该会话正在同步另一端完成的回合，请稍后重试",
       targetSessionMissing: "目标会话不存在",
       replyContentEmpty: "回复内容为空",
