+++ pinvou3-app/src/platform/web/bridge.js
@@ -335,7 +335,6 @@
       turnAlreadyInProgress: "⚠️ This chat is already processing a turn. The duplicate send was not executed.",
       compactStart: "⏳ Compacting context", compactDone: "✓ Context compacted", compactFail: "⚠️ Compaction failed", compactAuto: " (auto)",
       compactPruneMerged: "Auto-compaction: tool-result cleanup, messages unchanged",
-      shellCleanupFailed: "⚠️ Some background tasks could not be stopped. You can stop them individually from the background task list.",
       gpuUnavailable: "GPU info unavailable",
       superOn: "⚠️ Super permission enabled", superOff: "Super permission disabled",
       approved: "✅ Approved", echoGo: "✅ Do it",
@@ -447,7 +446,6 @@
       turnAlreadyInProgress: "⚠️ このチャットでは別のターンを処理中です。重複した送信は実行されませんでした。",
       compactStart: "⏳ コンテキストを圧縮中", compactDone: "✓ コンテキスト圧縮完了", compactFail: "⚠️ 圧縮に失敗", compactAuto: "（自動）",
       compactPruneMerged: "自動圧縮: ツール結果を整理、メッセージ数は不変",
-      shellCleanupFailed: "⚠️ 一部のバックグラウンドタスクを停止できませんでした。バックグラウンドタスク一覧から個別に停止できます。",
       gpuUnavailable: "GPU 情報を取得できません",
       superOn: "⚠️ スーパー権限が有効になりました", superOff: "スーパー権限が無効になりました",
       approved: "✅ 承認済み", echoGo: "✅ これでいく",
@@ -559,7 +557,6 @@
       turnAlreadyInProgress: "⚠️ 当前会话已有一轮正在处理，本次重复发送未执行。",
       compactStart: "⏳ 正在压缩上下文", compactDone: "✓ 上下文压缩完成", compactFail: "⚠️ 压缩失败", compactAuto: "（自动）",
       compactPruneMerged: "自动压缩：已整理工具结果，消息数不变",
-      shellCleanupFailed: "⚠️ 部分后台任务未能停止，可在后台任务列表中逐个停止。",
       gpuUnavailable: "GPU 信息不可用",
       superOn: "⚠️ 超级权限已开启", superOff: "超级权限已关闭",
       approved: "✅ 已批准", echoGo: "✅ 就这么干",
@@ -4638,26 +4635,6 @@
     }]);
   }
 
-  function recordTurnCompleted(payload) {
-    var openStart = latestOpenTimelineStart();
-    var turnId = state.activeTurnTimelineId || (openStart && openStart.turn_id);
-    if (!turnId) return;
-    var timestamp = Date.now();
-    var start = openStart || (state.turnTimeline || []).find(function (event) {
-      return event && event.turn_id === turnId && event.event === "user_start";
-    });
-    state.turnTimeline = (state.turnTimeline || []).concat([{
-      turn_id: turnId,
-      event: "assistant_done",
-      timestamp: timestamp,
-      ts: new Date(timestamp).toISOString(),
-      status: payload && payload.status || (payload && payload.error ? "Failed" : "Completed"),
-      error: payload && payload.error || null,
-      ui_turn_index: start && start.ui_turn_index,
-    }]);
-    state.activeTurnTimelineId = null;
-  }
-
   function preserveInterruptedAssistantPresentation() {
     var userItemIndex = -1;
     var afterMessageIndex = -1;
@@ -5084,7 +5061,11 @@
     runSyncOnSession(sid, function () {
       if (isScheduledRunSession(sid)) markScheduledInitialTurnTerminal(sid);
       var error = e.payload && e.payload.error;
-      recordTurnCompleted(e.payload || {});
+      window.PinvouWebTurnTerminal.recordCompleted(
+        state,
+        latestOpenTimelineStart(),
+        e.payload || {}
+      );
       // 401/鉴权失败:刷新 effectiveModelConfig → 前端拦截遮罩自动弹出引导配置。
       // \b401\b 词边界锚定,避免误匹配 "port 4014"/"row 401" 等含 401 子串的无关报错。
       if (error && /\b401\b|unauthorized|authentication/i.test(String(error))) loadEffectiveModelConfig();
@@ -5102,20 +5083,7 @@
           });
         }
       }
-      if (e.payload && e.payload.shell_cleanup_failed) {
-        var cleanupNotice = bt("shellCleanupFailed");
-        var cleanupNoticeItem = state.chatItems.find(function (item) {
-          return item && item.turnErrorNotice && item.text === cleanupNotice;
-        });
-        if (cleanupNoticeItem) {
-          cleanupNoticeItem.legacyConversationOnly = true;
-        } else {
-          addSystemItem(cleanupNotice, {
-            turnErrorNotice: true,
-            legacyConversationOnly: true,
-          });
-        }
-      }
+      window.PinvouBridgeMessages.showShellCleanupFailure(e.payload, state, addSystemItem);
       var terminalStatus = String(e.payload && e.payload.status || "").toLowerCase();
       var interrupted = terminalStatus === "interrupted" ||
         terminalStatus === "cancelled" || terminalStatus === "canceled";
