+++ pinvou3-app/src/shared/i18n.js
@@ -479,7 +479,6 @@ const dict = {
         hintSlow30: isLocal => isLocal ? '本地模型生成中' : '生成中…', hintSlow120: isLocal => isLocal ? '本地模型在生成长内容,请耐心等' : '正在生成长内容,请耐心等',
         artifactLabel: '成品', clickToOpen: '点击打开 →',
         toolRunning: '运行中...', toolDone: '完成', toolFailed: '失败', shellCancelFailed: '取消失败',
-        shellCleanupFailed: '部分后台任务未能停止，可在后台任务列表中逐个停止。',
         tsDone: '已完成', tsInProgress: '进行中', tsPending: '待办', tsLine: '行',
         receiptNote: '⋯ 输出较长，仅显示开头', receiptEmpty: '（输出较长，未展开）',
         listDirCount: n => `${n} 个条目`,
@@ -796,7 +795,6 @@ const dict = {
         hintSlow30: isLocal => isLocal ? 'local model generating' : 'generating…', hintSlow120: isLocal => isLocal ? 'local model is generating long content, hang tight' : 'generating long content, hang tight',
         artifactLabel: 'Artifact', clickToOpen: 'Open →',
         toolRunning: 'Running...', toolDone: 'Done', toolFailed: 'Failed', shellCancelFailed: 'Cancel failed',
-        shellCleanupFailed: 'Some background tasks could not be stopped. You can stop them individually from the background task list.',
         tsDone: 'completed', tsInProgress: 'in progress', tsPending: 'pending', tsLine: 'L',
         receiptNote: '⋯ Long output — showing the beginning only', receiptEmpty: '(long output, not expanded)',
         listDirCount: n => `${n} entries`,
@@ -1112,7 +1110,6 @@ const dict = {
         hintSlow30: isLocal => isLocal ? 'ローカルモデルが生成中' : '生成中…', hintSlow120: isLocal => isLocal ? 'ローカルモデルが長い内容を生成中です。しばらくお待ちください' : '長い内容を生成中です。しばらくお待ちください',
         artifactLabel: '成果物', clickToOpen: 'クリックで開く →',
         toolRunning: '実行中...', toolDone: '完了', toolFailed: '失敗', shellCancelFailed: 'キャンセルに失敗しました',
-        shellCleanupFailed: '一部のバックグラウンドタスクを停止できませんでした。バックグラウンドタスク一覧から個別に停止できます。',
         tsDone: '完了', tsInProgress: '進行中', tsPending: '未着手', tsLine: '行',
         receiptNote: '⋯ 出力が長いため冒頭のみ表示', receiptEmpty: '（出力が長いため省略）',
         listDirCount: n => `${n} 件`,
